refactor: split up code from `predict_utils`
