+++ src/syn_net/utils/predict_utils.py
@@ -2,38 +2,25 @@
 This file contains various utils for creating molecular embeddings and for
 decoding synthetic trees.
 """
+from typing import Callable, Tuple
+
 import numpy as np
+import pytorch_lightning as pl
 import rdkit
-from tqdm import tqdm
 import torch
 from rdkit import Chem
-from rdkit import DataStructs
-from rdkit.Chem import AllChem
 from sklearn.neighbors import BallTree
-from dgl.nn.pytorch.glob import AvgPooling
-from dgl.nn.pytorch.glob import AvgPooling
-from dgllife.model import load_pretrained
-from dgllife.utils import mol_to_bigraph, PretrainAtomFeaturizer, PretrainBondFeaturizer
+from syn_net.encoders.distances import cosine_distance, tanimoto_similarity
+from syn_net.encoders.fingerprints import mol_fp
+from syn_net.encoders.utils import one_hot_encoder
+from syn_net.utils.data_utils import Reaction, SyntheticTree
 from tdc.chem_utils import MolConvert
-from syn_net.models.mlp import MLP
-from syn_net.utils.data_utils import SyntheticTree
-import functools
 
 # create a random seed for NumPy
 np.random.seed(6)
 
-
-@functools.lru_cache(1)
-def _fetch_gin_pretrained_model(model_name: str):
-    """Get a GIN pretrained model to use for creating molecular embeddings"""
-    device = 'cpu'
-    model = load_pretrained(model_name).to(device) # used to learn embedding
-    model.eval()
-    return model
-
-
 # general functions
-def can_react(state, rxns):
+def can_react(state, rxns: list[Reaction]) -> Tuple[int, list[bool]]:
     """
     Determines if two molecules can react using any of the input reactions.
 
@@ -52,7 +39,8 @@ def can_react(state, rxns):
     reaction_mask = [int(rxn.run_reaction([mol1, mol2]) is not None) for rxn in rxns]
     return sum(reaction_mask), reaction_mask
 
-def get_action_mask(state, rxns):
+
+def get_action_mask(state: list, rxns: list[Reaction]) -> np.ndarray:
     """
     Determines which actions can apply to a given state in the synthetic tree
     and returns a mask for which actions can apply.
@@ -71,19 +59,21 @@ def get_action_mask(state, rxns):
     """
     # Action: (Add: 0, Expand: 1, Merge: 2, End: 3)
     if len(state) == 0:
-        return np.array([1, 0, 0, 0])
+        mask = [1, 0, 0, 0]
     elif len(state) == 1:
-        return np.array([1, 1, 0, 1])
+        mask = [1, 1, 0, 1]
     elif len(state) == 2:
         can_react_, _ = can_react(state, rxns)
         if can_react_:
-            return np.array([0, 1, 1, 0])
+            mask = [0, 1, 1, 0]
         else:
-            return np.array([0, 1, 0, 0])
+            mask = [0, 1, 0, 0]
     else:
-        raise ValueError('Problem with state.')
+        raise ValueError("Problem with state.")
+    return np.asarray(mask, dtype=bool)
 
-def get_reaction_mask(smi, rxns):
+
+def get_reaction_mask(smi: str, rxns: list[Reaction]):
     """
     Determines which reaction templates can apply to the input molecule.
 
@@ -109,6 +99,7 @@ def get_reaction_mask(smi, rxns):
 
     if sum(reaction_mask) == 0:
         return None, None
+
     available_list = []
     mol = rdkit.Chem.MolFromSmiles(smi)
     for i, rxn in enumerate(rxns):
@@ -119,7 +110,7 @@ def get_reaction_mask(smi, rxns):
             elif rxn.is_reactant_second(mol):
                 available_list.append(rxn.available_reactants[0])
             else:
-                raise ValueError('Check the reactants')
+                raise ValueError("Check the reactants")
 
             if len(available_list[-1]) == 0:
                 reaction_mask[i] = 0
@@ -129,170 +120,8 @@ def get_reaction_mask(smi, rxns):
 
     return reaction_mask, available_list
 
-def graph_construction_and_featurization(smiles):
-    """
-    Constructs graphs from SMILES and featurizes them.
 
-    Args:
-        smiles (list of str): Contains SMILES of molecules to embed.
-
-    Returns:
-        graphs (list of DGLGraph): List of graphs constructed and featurized.
-        success (list of bool): Indicators for whether the SMILES string can be
-            parsed by RDKit.
-    """
-    graphs = []
-    success = []
-    for smi in tqdm(smiles):
-        try:
-            mol = Chem.MolFromSmiles(smi)
-            if mol is None:
-                success.append(False)
-                continue
-            g = mol_to_bigraph(mol, add_self_loop=True,
-                                node_featurizer=PretrainAtomFeaturizer(),
-                                edge_featurizer=PretrainBondFeaturizer(),
-                                canonical_atom_order=False)
-            graphs.append(g)
-            success.append(True)
-        except:
-            success.append(False)
-
-    return graphs, success
-
-def one_hot_encoder(dim, space):
-    """
-    Create a one-hot encoded vector of length=`space`, with a non-zero element
-    at the index given by `dim`.
-
-    Args:
-        dim (int): Non-zero bit in one-hot vector.
-        space (int): Length of one-hot encoded vector.
-
-    Returns:
-        vec (np.ndarray): One-hot encoded vector.
-    """
-    vec = np.zeros((1, space))
-    vec[0, dim] = 1
-    return vec
-
-def mol_embedding(smi, device='cpu', readout=AvgPooling()):
-    """
-    Constructs a graph embedding using the GIN network for an input SMILES.
-
-    Args:
-        smi (str): A SMILES string.
-        device (str): Indicates the device to run on ('cpu' or 'cuda:0'). Default 'cpu'.
-
-    Returns:
-        np.ndarray: Either a zeros array or the graph embedding.
-    """
-    name = 'gin_supervised_contextpred'
-    gin_pretrained_model = _fetch_gin_pretrained_model(name)
-
-    # get the embedding
-    if smi is None:
-        return np.zeros(300)
-    else:
-        mol = Chem.MolFromSmiles(smi)
-        # convert RDKit.Mol into featurized bi-directed DGLGraph
-        g = mol_to_bigraph(mol, add_self_loop=True,
-                           node_featurizer=PretrainAtomFeaturizer(),
-                           edge_featurizer=PretrainBondFeaturizer(),
-                           canonical_atom_order=False)
-        bg = g.to(device)
-        nfeats = [bg.ndata.pop('atomic_number').to(device),
-                  bg.ndata.pop('chirality_type').to(device)]
-        efeats = [bg.edata.pop('bond_type').to(device),
-                  bg.edata.pop('bond_direction_type').to(device)]
-        with torch.no_grad():
-            node_repr = gin_pretrained_model(bg, nfeats, efeats)
-        return readout(bg, node_repr).detach().cpu().numpy().reshape(-1, ).tolist()
-
-
-def get_mol_embedding(smi, model, device='cpu', readout=AvgPooling()):
-    """
-    Computes the molecular graph embedding for the input SMILES.
-
-    Args:
-        smi (str): SMILES of molecule to embed.
-        model (dgllife.model, optional): Pre-trained NN model to use for
-            computing the embedding.
-        device (str, optional): Indicates the device to run on. Defaults to 'cpu'.
-        readout (dgl.nn.pytorch.glob, optional): Readout function to use for
-            computing the graph embedding. Defaults to readout.
-
-    Returns:
-        torch.Tensor: Learned embedding for the input molecule.
-    """
-    mol = Chem.MolFromSmiles(smi)
-    g = mol_to_bigraph(mol, add_self_loop=True,
-                       node_featurizer=PretrainAtomFeaturizer(),
-                       edge_featurizer=PretrainBondFeaturizer(),
-                       canonical_atom_order=False)
-    bg = g.to(device)
-    nfeats = [bg.ndata.pop('atomic_number').to(device),
-              bg.ndata.pop('chirality_type').to(device)]
-    efeats = [bg.edata.pop('bond_type').to(device),
-              bg.edata.pop('bond_direction_type').to(device)]
-    with torch.no_grad():
-        node_repr = model(bg, nfeats, efeats)
-    return readout(bg, node_repr).detach().cpu().numpy()[0]
-
-def mol_fp(smi, _radius=2, _nBits=4096):
-    """
-    Computes the Morgan fingerprint for the input SMILES.
-
-    Args:
-        smi (str): SMILES for molecule to compute fingerprint for.
-        _radius (int, optional): Fingerprint radius to use. Defaults to 2.
-        _nBits (int, optional): Length of fingerprint. Defaults to 1024.
-
-    Returns:
-        features (np.ndarray): For valid SMILES, this is the fingerprint.
-            Otherwise, if the input SMILES is bad, this will be a zero vector.
-    """
-    if smi is None:
-        return np.zeros(_nBits)
-    else:
-        mol = Chem.MolFromSmiles(smi)
-        features_vec = AllChem.GetMorganFingerprintAsBitVect(mol, _radius, _nBits)
-        return np.array(features_vec)
-
-def cosine_distance(v1, v2, eps=1e-15):
-    """
-    Computes the cosine similarity between two vectors.
-
-    Args:
-        v1 (np.ndarray): First vector.
-        v2 (np.ndarray): Second vector.
-        eps (float, optional): Small value, for numerical stability. Defaults
-            to 1e-15.
-
-    Returns:
-        float: The cosine similarity.
-    """
-    return (1 - np.dot(v1, v2)
-            / (np.linalg.norm(v1, ord=2) * np.linalg.norm(v2, ord=2) + eps))
-
-def ce_distance(y, y_pred, eps=1e-15):
-    """
-    Computes the cross-entropy between two vectors.
-
-    Args:
-        y (np.ndarray): First vector.
-        y_pred (np.ndarray): Second vector.
-        eps (float, optional): Small value, for numerical stability. Defaults
-            to 1e-15.
-
-    Returns:
-        float: The cross-entropy.
-    """
-    y_pred = np.clip(y_pred, eps, 1 - eps)
-    return - np.sum((y * np.log(y_pred) + (1 - y) * np.log(1 - y_pred)))
-
-
-def nn_search(_e, _tree, _k=1):
+def nn_search(_e: np.ndarray, _tree: BallTree, _k: int = 1) -> Tuple[float, float]:
     """
     Conducts a nearest neighbor search to find the molecule from the tree most
     simimilar to the input embedding.
@@ -310,80 +139,57 @@ def nn_search(_e, _tree, _k=1):
     dist, ind = _tree.query(_e, k=_k)
     return dist[0][0], ind[0][0]
 
-def graph_construction_and_featurization(smiles):
-    """
-    Constructs graphs from SMILES and featurizes them.
 
-    Args:
-        smiles (list of str): SMILES of molecules, for embedding computation.
+def nn_search_rt1(_e: np.ndarray, _tree: BallTree, _k: int = 1) -> Tuple[np.ndarray, np.ndarray]:
+    dist, ind = _tree.query(_e, k=_k)
+    return dist[0], ind[0]
 
-    Returns:
-        graphs (list of DGLGraph): List of graphs constructed and featurized.
-        success (list of bool): Indicators for whether the SMILES string can be
-            parsed by RDKit.
-    """
-    graphs  = []
-    success = []
-    for smi in tqdm(smiles):
-        try:
-            mol = Chem.MolFromSmiles(smi)
-            if mol is None:
-                success.append(False)
-                continue
-            g = mol_to_bigraph(mol, add_self_loop=True,
-                               node_featurizer=PretrainAtomFeaturizer(),
-                               edge_featurizer=PretrainBondFeaturizer(),
-                               canonical_atom_order=False)
-            graphs.append(g)
-            success.append(True)
-        except:
-            success.append(False)
 
-    return graphs, success
-
-def set_embedding(z_target, state, nbits, _mol_embedding=get_mol_embedding):
+def set_embedding(z_target: np.ndarray, state: list[str], nbits: int, _mol_embedding: Callable):
     """
     Computes embeddings for all molecules in the input space.
+    Embedding = [z_mol1, z_mol2, z_target]
 
     Args:
         z_target (np.ndarray): Embedding for the target molecule.
-        state (list): Contains molecules in the current state, if not the
-            initial state.
+        state (list): Contains molecules in the current state, if not the initial state.
         nbits (int): Length of fingerprint.
-        _mol_embedding (Callable, optional): Function to use for computing the
-            embeddings of the first and second molecules in the state. Defaults
-            to `get_mol_embedding`.
+        _mol_embedding (Callable): Function to use for computing the
+            embeddings of the first and second molecules in the state.
 
     Returns:
         np.ndarray: Embedding consisting of the concatenation of the target
             molecule with the current molecules (if available) in the input state.
     """
     if len(state) == 0:
-        return np.concatenate([np.zeros((1, 2 * nbits)), z_target], axis=1)
+        embedding =  np.concatenate([np.zeros((1, 2 * nbits)), z_target], axis=1)
     else:
         e1 = np.expand_dims(_mol_embedding(state[0]), axis=0)
         if len(state) == 1:
             e2 = np.zeros((1, nbits))
         else:
             e2 = _mol_embedding(state[1])
-        return np.concatenate([e1, e2, z_target], axis=1)
+        embedding = np.concatenate([e1, e2, z_target], axis=1)
+    return embedding
 
-def synthetic_tree_decoder(z_target,
-                           building_blocks,
-                           bb_dict,
-                           reaction_templates,
-                           mol_embedder,
-                           action_net,
-                           reactant1_net,
-                           rxn_net,
-                           reactant2_net,
-                           bb_emb,
-                           rxn_template,
-                           n_bits,
-                           max_step=15):
+def synthetic_tree_decoder(
+    z_target: np.ndarray,
+    building_blocks: list[str],
+    bb_dict: dict[str, int],
+    reaction_templates: list[Reaction],
+    mol_embedder,
+    action_net: pl.LightningModule,
+    reactant1_net: pl.LightningModule,
+    rxn_net: pl.LightningModule,
+    reactant2_net: pl.LightningModule,
+    bb_emb: np.ndarray,
+    rxn_template: str,
+    n_bits: int,
+    max_step: int = 15,
+) -> Tuple[SyntheticTree, int]:
     """
-    Computes the synthetic tree given an input molecule embedding, using the
-    Action, Reaction, Reactant1, and Reactant2 networks and a greedy search
+    Computes a synthetic tree given an input molecule embedding.
+    Uses the Action, Reaction, Reactant1, and Reactant2 networks and a greedy search.
 
     Args:
         z_target (np.ndarray): Embedding for the target molecule
@@ -408,15 +214,15 @@ def synthetic_tree_decoder(z_target,
             terminated).
     """
     # Initialization
-    tree       = SyntheticTree()
-    kdtree     = BallTree(bb_emb, metric=cosine_distance)
+    tree = SyntheticTree()
     mol_recent = None
+    kdtree = BallTree(bb_emb, metric=cosine_distance)  # TODO: cache this or use class
+    z_target = np.atleast_2d(z_target)
 
     # Start iteration
-    # try:
     for i in range(max_step):
         # Encode current state
-        state = tree.get_state() # a set
+        state = tree.get_state()  # a list
         z_state = set_embedding(z_target, state, nbits=n_bits, _mol_embedding=mol_fp)
 
         # Predict action type, masked selection
@@ -426,38 +232,39 @@ def synthetic_tree_decoder(z_target,
         action_mask = get_action_mask(tree.get_state(), reaction_templates)
         act = np.argmax(action_proba * action_mask)
 
+        # Continue growing tree?
+        if act == 3:  # End
+            break
+
         z_mol1 = reactant1_net(torch.Tensor(z_state))
         z_mol1 = z_mol1.detach().numpy()
 
         # Select first molecule
-        if act == 3:
-            # End
-            break
-        elif act == 0:
+        if act == 0:
             # Add
             dist, ind = nn_search(z_mol1, _tree=kdtree)
             mol1 = building_blocks[ind]
-        else:
+        elif act == 1 or act == 2:
             # Expand or Merge
             mol1 = mol_recent
+        else:
+            raise ValueError(f"Unexpected action {act}.")
 
         z_mol1 = mol_fp(mol1)
+        z_mol1 = np.atleast_2d(z_mol1)  # (1,4096)
 
         # Select reaction
-        try:
-            reaction_proba = rxn_net(torch.Tensor(np.concatenate([z_state, z_mol1], axis=1)))
-        except:
-            z_mol1 = np.expand_dims(z_mol1, axis=0)
-            reaction_proba = rxn_net(torch.Tensor(np.concatenate([z_state, z_mol1], axis=1)))
-        reaction_proba = reaction_proba.squeeze().detach().numpy() + 1e-10
+        z = np.concatenate([z_state, z_mol1], axis=1)
+        reaction_proba = rxn_net(torch.Tensor(z))
+        reaction_proba = reaction_proba.squeeze().detach().numpy() + 1e-10  # (nReactionTemplate)
 
-        if act != 2:
-            reaction_mask, available_list = get_reaction_mask(smi=mol1,
-                                                              rxns=reaction_templates)
-        else:
+        if act != 2:  # add or expand
+            reaction_mask, available_list = get_reaction_mask(mol1, reaction_templates)
+        else:  # merge
             _, reaction_mask = can_react(tree.get_state(), reaction_templates)
             available_list = [[] for rxn in reaction_templates]
 
+        # If we ended up in a state where no reaction is possible, end this iteration.
         if reaction_mask is None:
             if len(state) == 1:
                 act = 3
@@ -465,33 +272,32 @@ def synthetic_tree_decoder(z_target,
             else:
                 break
 
+        # Select reaction template
         rxn_id = np.argmax(reaction_proba * reaction_mask)
         rxn = reaction_templates[rxn_id]
 
+        NUMBER_OF_REACTION_TEMPLATES = {
+            "hb": 91,
+            "pis": 4700,
+            "unittest": 3,
+        }  # TODO: Refactor / use class
+
+        # Select 2nd reactant
         if rxn.num_reactant == 2:
-            # Select second molecule
-            if act == 2:
-                # Merge
+            if act == 2:  # Merge
                 temp = set(state) - set([mol1])
                 mol2 = temp.pop()
-            else:
-                # Add or Expand
-                if rxn_template == 'hb':
-                    num_rxns = 91
-                elif rxn_template == 'pis':
-                    num_rxns = 4700
-                else:
-                    num_rxns = 3  # unit testing uses only 3 reaction templates
-                reactant2_net_input = torch.Tensor(
-                    np.concatenate([z_state, z_mol1, one_hot_encoder(rxn_id, num_rxns)],
-                                   axis=1)
-                )
-                z_mol2 = reactant2_net(reactant2_net_input)
+            else:  # Add or Expand
+                x_rxn = one_hot_encoder(rxn_id, NUMBER_OF_REACTION_TEMPLATES[rxn_template])
+                x_rct2 = np.concatenate([z_state, z_mol1, x_rxn], axis=1)
+                z_mol2 = reactant2_net(torch.Tensor(x_rct2))
                 z_mol2 = z_mol2.detach().numpy()
-                available = available_list[rxn_id]
-                available = [bb_dict[available[i]] for i in range(len(available))]
+                available = available_list[rxn_id]  # list[str], list of reactants for this rxn
+                available = [bb_dict[available[i]] for i in range(len(available))]  # list[int]
                 temp_emb = bb_emb[available]
-                available_tree = BallTree(temp_emb, metric=cosine_distance)
+                available_tree = BallTree(
+                    temp_emb, metric=cosine_distance
+                )  # TODO: evaluate if distance matrix is faster/feasible as this BallTree is discarded immediately.
                 dist, ind = nn_search(z_mol2, _tree=available_tree)
                 mol2 = building_blocks[available[ind]]
         else:
@@ -517,300 +323,23 @@ def synthetic_tree_decoder(z_target,
 
     return tree, act
 
-def load_modules_from_checkpoint(path_to_act, path_to_rt1, path_to_rxn, path_to_rt2, featurize, rxn_template, out_dim, nbits, ncpu):
 
-    if rxn_template == 'unittest':
-
-        act_net = MLP.load_from_checkpoint(path_to_act,
-                                           input_dim=int(3 * nbits),
-                                           output_dim=4,
-                                           hidden_dim=100,
-                                           num_layers=3,
-                                           dropout=0.5,
-                                           num_dropout_layers=1,
-                                           task='classification',
-                                           loss='cross_entropy',
-                                           valid_loss='accuracy',
-                                           optimizer='adam',
-                                           learning_rate=1e-4,
-                                           ncpu=ncpu)
-
-        rt1_net = MLP.load_from_checkpoint(path_to_rt1,
-                                           input_dim=int(3 * nbits),
-                                           output_dim=out_dim,
-                                           hidden_dim=100,
-                                           num_layers=3,
-                                           dropout=0.5,
-                                           num_dropout_layers=1,
-                                           task='regression',
-                                           loss='mse',
-                                           valid_loss='mse',
-                                           optimizer='adam',
-                                           learning_rate=1e-4,
-                                           ncpu=ncpu)
-
-        rxn_net = MLP.load_from_checkpoint(path_to_rxn,
-                                           input_dim=int(4 * nbits),
-                                           output_dim=3,
-                                           hidden_dim=100,
-                                           num_layers=5,
-                                           dropout=0.5,
-                                           num_dropout_layers=1,
-                                           task='classification',
-                                           loss='cross_entropy',
-                                           valid_loss='accuracy',
-                                           optimizer='adam',
-                                           learning_rate=1e-4,
-                                           ncpu=ncpu)
-
-        rt2_net = MLP.load_from_checkpoint(path_to_rt2,
-                                           input_dim=int(4 * nbits + 3),
-                                           output_dim=out_dim,
-                                           hidden_dim=100,
-                                           num_layers=3,
-                                           dropout=0.5,
-                                           num_dropout_layers=1,
-                                           task='regression',
-                                           loss='mse',
-                                           valid_loss='mse',
-                                           optimizer='adam',
-                                           learning_rate=1e-4,
-                                           ncpu=ncpu)
-    elif featurize == 'fp':
-
-        act_net = MLP.load_from_checkpoint(path_to_act,
-                                           input_dim=int(3 * nbits),
-                                           output_dim=4,
-                                           hidden_dim=1000,
-                                           num_layers=5,
-                                           dropout=0.5,
-                                           num_dropout_layers=1,
-                                           task='classification',
-                                           loss='cross_entropy',
-                                           valid_loss='accuracy',
-                                           optimizer='adam',
-                                           learning_rate=1e-4,
-                                           ncpu=ncpu)
-
-        rt1_net = MLP.load_from_checkpoint(path_to_rt1,
-                                           input_dim=int(3 * nbits),
-                                           output_dim=int(out_dim),
-                                           hidden_dim=1200,
-                                           num_layers=5,
-                                           dropout=0.5,
-                                           num_dropout_layers=1,
-                                           task='regression',
-                                           loss='mse',
-                                           valid_loss='mse',
-                                           optimizer='adam',
-                                           learning_rate=1e-4,
-                                           ncpu=ncpu)
-
-        if rxn_template == 'hb':
-
-            rxn_net = MLP.load_from_checkpoint(path_to_rxn,
-                                               input_dim=int(4 * nbits),
-                                               output_dim=91,
-                                               hidden_dim=3000,
-                                               num_layers=5,
-                                               dropout=0.5,
-                                               num_dropout_layers=1,
-                                               task='classification',
-                                               loss='cross_entropy',
-                                               valid_loss='accuracy',
-                                               optimizer='adam',
-                                               learning_rate=1e-4,
-                                               ncpu=ncpu)
-
-            rt2_net = MLP.load_from_checkpoint(path_to_rt2,
-                                               input_dim=int(4 * nbits + 91),
-                                               output_dim=int(out_dim),
-                                               hidden_dim=3000,
-                                               num_layers=5,
-                                               dropout=0.5,
-                                               num_dropout_layers=1,
-                                               task='regression',
-                                               loss='mse',
-                                               valid_loss='mse',
-                                               optimizer='adam',
-                                               learning_rate=1e-4,
-                                               ncpu=ncpu)
-
-        elif rxn_template == 'pis':
-
-            rxn_net = MLP.load_from_checkpoint(path_to_rxn,
-                                               input_dim=int(4 * nbits),
-                                               output_dim=4700,
-                                               hidden_dim=4500,
-                                               num_layers=5,
-                                               dropout=0.5,
-                                               num_dropout_layers=1,
-                                               task='classification',
-                                               loss='cross_entropy',
-                                               valid_loss='accuracy',
-                                               optimizer='adam',
-                                               learning_rate=1e-4,
-                                               ncpu=ncpu)
-
-            rt2_net = MLP.load_from_checkpoint(path_to_rt2,
-                                               input_dim=int(4 * nbits + 4700),
-                                               output_dim=out_dim,
-                                               hidden_dim=3000,
-                                               num_layers=5,
-                                               dropout=0.5,
-                                               num_dropout_layers=1,
-                                               task='regression',
-                                               loss='mse',
-                                               valid_loss='mse',
-                                               optimizer='adam',
-                                               learning_rate=1e-4,
-                                               ncpu=ncpu)
-
-    elif featurize == 'gin':
-
-        act_net = MLP.load_from_checkpoint(path_to_act,
-                                           input_dim=int(2 * nbits + out_dim),
-                                           output_dim=4,
-                                           hidden_dim=1000,
-                                           num_layers=5,
-                                           dropout=0.5,
-                                           num_dropout_layers=1,
-                                           task='classification',
-                                           loss='cross_entropy',
-                                           valid_loss='accuracy',
-                                           optimizer='adam',
-                                           learning_rate=1e-4,
-                                           ncpu=ncpu)
-
-        rt1_net = MLP.load_from_checkpoint(path_to_rt1,
-                                           input_dim=int(2 * nbits + out_dim),
-                                           output_dim=out_dim,
-                                           hidden_dim=1200,
-                                           num_layers=5,
-                                           dropout=0.5,
-                                           num_dropout_layers=1,
-                                           task='regression',
-                                           loss='mse',
-                                           valid_loss='mse',
-                                           optimizer='adam',
-                                           learning_rate=1e-4,
-                                           ncpu=ncpu)
-
-        if rxn_template == 'hb':
-
-            rxn_net = MLP.load_from_checkpoint(path_to_rxn,
-                                               input_dim=int(3 * nbits + out_dim),
-                                               output_dim=91,
-                                               hidden_dim=3000,
-                                               num_layers=5,
-                                               dropout=0.5,
-                                               num_dropout_layers=1,
-                                               task='classification',
-                                               loss='cross_entropy',
-                                               valid_loss='accuracy',
-                                               optimizer='adam',
-                                               learning_rate=1e-4,
-                                               ncpu=ncpu)
-
-            rt2_net = MLP.load_from_checkpoint(path_to_rt2,
-                                               input_dim=int(3 * nbits + out_dim + 91),
-                                               output_dim=out_dim,
-                                               hidden_dim=3000,
-                                               num_layers=5,
-                                               dropout=0.5,
-                                               num_dropout_layers=1,
-                                               task='regression',
-                                               loss='mse',
-                                               valid_loss='mse',
-                                               optimizer='adam',
-                                               learning_rate=1e-4,
-                                               ncpu=ncpu)
-
-        elif rxn_template == 'pis':
-
-            rxn_net = MLP.load_from_checkpoint(path_to_rxn,
-                                               input_dim=int(3 * nbits + out_dim),
-                                               output_dim=4700,
-                                               hidden_dim=3000,
-                                               num_layers=5,
-                                               dropout=0.5,
-                                               num_dropout_layers=1,
-                                               task='classification',
-                                               loss='cross_entropy',
-                                               valid_loss='accuracy',
-                                               optimizer='adam',
-                                               learning_rate=1e-4,
-                                               ncpu=ncpu)
-
-            rt2_net = MLP.load_from_checkpoint(path_to_rt2,
-                                               input_dim=int(3 * nbits + out_dim + 4700),
-                                               output_dim=out_dim,
-                                               hidden_dim=3000,
-                                               num_layers=5,
-                                               dropout=0.5,
-                                               num_dropout_layers=1,
-                                               task='regression',
-                                               loss='mse',
-                                               valid_loss='mse',
-                                               optimizer='adam',
-                                               learning_rate=1e-4,
-                                               ncpu=ncpu)
-
-    act_net.eval()
-    rt1_net.eval()
-    rxn_net.eval()
-    rt2_net.eval()
-
-    return act_net, rt1_net, rxn_net, rt2_net
-
-def _tanimoto_similarity(fp1, fp2):
-    """
-    Returns the Tanimoto similarity between two molecular fingerprints.
-
-    Args:
-        fp1 (np.ndarray): Molecular fingerprint 1.
-        fp2 (np.ndarray): Molecular fingerprint 2.
-
-    Returns:
-        np.float: Tanimoto similarity.
-    """
-    return np.sum(fp1 * fp2) / (np.sum(fp1) + np.sum(fp2) - np.sum(fp1 * fp2))
-
-def tanimoto_similarity(target_fp, smis):
-    """
-    Returns the Tanimoto similarities between a target fingerprint and molecules
-    in an input list of SMILES.
-
-    Args:
-        target_fp (np.ndarray): Contains the reference (target) fingerprint.
-        smis (list of str): Contains SMILES to compute similarity to.
-
-    Returns:
-        list of np.ndarray: Contains Tanimoto similarities.
-    """
-    fps = [mol_fp(smi, 2, 4096) for smi in smis]
-    return [_tanimoto_similarity(target_fp, fp) for fp in fps]
-
-
-# functions used in the *_multireactant.py
-def nn_search_rt1(_e, _tree, _k=1):
-    dist, ind = _tree.query(_e, k=_k)
-    return dist[0], ind[0]
-
-def synthetic_tree_decoder_rt1(z_target,
-                                building_blocks,
-                                bb_dict,
-                                reaction_templates,
-                                mol_embedder,
-                                action_net,
-                                reactant1_net,
-                                rxn_net,
-                                reactant2_net,
-                                bb_emb,
-                                rxn_template,
-                                n_bits,
-                                max_step=15,
-                                rt1_index=0):
+def synthetic_tree_decoder_rt1(
+    z_target: np.ndarray,
+    building_blocks: list[str],
+    bb_dict: dict[str, int],
+    reaction_templates: list[Reaction],
+    mol_embedder,
+    action_net: pl.LightningModule,
+    reactant1_net: pl.LightningModule,
+    rxn_net: pl.LightningModule,
+    reactant2_net: pl.LightningModule,
+    bb_emb: np.ndarray,
+    rxn_template: str,
+    n_bits: int,
+    max_step: int = 15,
+    rt1_index=0,
+) -> Tuple[SyntheticTree, int]:
     """
     Computes the synthetic tree given an input molecule embedding, using the
     Action, Reaction, Reactant1, and Reactant2 networks and a greedy search.
@@ -835,7 +364,7 @@ def synthetic_tree_decoder_rt1(z_target,
             synthetic tree
         rt1_index (int, optional): Index for molecule in the building blocks
             corresponding to reactant 1.
-    
+
     Returns:
         tree (SyntheticTree): The final synthetic tree
         act (int): The final action (to know if the tree was "properly"
@@ -844,59 +373,61 @@ def synthetic_tree_decoder_rt1(z_target,
     # Initialization
     tree = SyntheticTree()
     mol_recent = None
-    kdtree = BallTree(bb_emb, metric=cosine_distance) # TODO: cache this or use class
-    z_target =  np.atleast_2d(z_target)
+    kdtree = BallTree(bb_emb, metric=cosine_distance)  # TODO: cache this or use class
+    z_target = np.atleast_2d(z_target)
     # Start iteration
     for i in range(max_step):
         # Encode current state
-        state = tree.get_state() # a list
+        state = tree.get_state()  # a list
         z_state = set_embedding(z_target, state, nbits=n_bits, _mol_embedding=mol_fp)
 
         # Predict action type, masked selection
         # Action: (Add: 0, Expand: 1, Merge: 2, End: 3)
-        action_proba = action_net(torch.Tensor(z_state)) # (1,4)
+        action_proba = action_net(torch.Tensor(z_state))  # (1,4)
         action_proba = action_proba.squeeze().detach().numpy() + 1e-10
-        action_mask  = get_action_mask(tree.get_state(), reaction_templates)
-        act          = np.argmax(action_proba * action_mask)
+        action_mask = get_action_mask(tree.get_state(), reaction_templates)
+        act = np.argmax(action_proba * action_mask)
 
-        # Continue growing tree? 
-        if act == 3: # End
+        # Continue growing tree?
+        if act == 3:  # End
             break
 
         z_mol1 = reactant1_net(torch.Tensor(z_state))
-        z_mol1 = z_mol1.detach().numpy() # (1,dimension_output_embedding), default: (1,256)
-
+        z_mol1 = z_mol1.detach().numpy()  # (1,dimension_output_embedding), default: (1,256)
 
         # Select first molecule
-        if act == 0: # Add
+        if act == 0:  # Add
             if mol_recent is not None:
                 dist, ind = nn_search(z_mol1)
                 mol1 = building_blocks[ind]
-            else: # no recent mol
-                dist, ind = nn_search_rt1(z_mol1, _tree=kdtree, _k=rt1_index+1) # TODO: why is there an option to select the k-th? rt1_index (???)
+            else:  # no recent mol
+                dist, ind = nn_search_rt1(
+                    z_mol1, _tree=kdtree, _k=rt1_index + 1
+                )  # TODO: why is there an option to select the k-th? rt1_index (???)
                 mol1 = building_blocks[ind[rt1_index]]
-        elif act==1 or act==2:
+        elif act == 1 or act == 2:
             # Expand or Merge
             mol1 = mol_recent
-        else:  
+        else:
             raise ValueError(f"Unexpected action {act}.")
 
-        z_mol1 = mol_fp(mol1) # (dimension_input_embedding=d), default (4096,)
-        z_mol1 = np.atleast_2d(z_mol1) # (1,4096)
+        z_mol1 = mol_fp(mol1)  # (dimension_input_embedding=d), default (4096,)
+        z_mol1 = np.atleast_2d(z_mol1)  # (1,4096)
 
         # Select reaction
-        z = np.concatenate([z_state, z_mol1], axis=1) # (1,4d)
+        z = np.concatenate([z_state, z_mol1], axis=1)  # (1,4d)
         reaction_proba = rxn_net(torch.Tensor(z))
-        reaction_proba = reaction_proba.squeeze().detach().numpy() + 1e-10 # (nReactionTemplate)
+        reaction_proba = reaction_proba.squeeze().detach().numpy() + 1e-10  # (nReactionTemplate)
 
-        if act != 2: # add or expand
+        if act != 2:  # add or expand
             reaction_mask, available_list = get_reaction_mask(mol1, reaction_templates)
-        else: # merge
+        else:  # merge
             _, reaction_mask = can_react(tree.get_state(), reaction_templates)
-            available_list = [[] for rxn in reaction_templates] # TODO: if act=merge, this is not used at all
+            available_list = [
+                [] for rxn in reaction_templates
+            ]  # TODO: if act=merge, this is not used at all
 
-        # If we ended up in a state where no reaction is possible, 
-        # end this iteration.
+        # If we ended up in a state where no reaction is possible, end this iteration.
         if reaction_mask is None:
             if len(state) == 1:
                 act = 3
@@ -906,30 +437,32 @@ def synthetic_tree_decoder_rt1(z_target,
 
         # Select reaction template
         rxn_id = np.argmax(reaction_proba * reaction_mask)
-        rxn    = reaction_templates[rxn_id]
+        rxn = reaction_templates[rxn_id]
 
         NUMBER_OF_REACTION_TEMPLATES = {
             "hb": 91,
             "pis": 4700,
             "unittest": 3,
-        } # TODO: Refactor / use class
+        }  # TODO: Refactor / use class
 
         # Select 2nd reactant
         if rxn.num_reactant == 2:
-            if act == 2: # Merge
+            if act == 2:  # Merge
                 temp = set(state) - set([mol1])
                 mol2 = temp.pop()
-            else: # Add or Expand
-                x_rxn = one_hot_encoder(rxn_id,NUMBER_OF_REACTION_TEMPLATES[rxn_template])
-                x_rct2 = np.concatenate([z_state,z_mol1, x_rxn],axis=1)
+            else:  # Add or Expand
+                x_rxn = one_hot_encoder(rxn_id, NUMBER_OF_REACTION_TEMPLATES[rxn_template])
+                x_rct2 = np.concatenate([z_state, z_mol1, x_rxn], axis=1)
                 z_mol2 = reactant2_net(torch.Tensor(x_rct2))
-                z_mol2         = z_mol2.detach().numpy()
-                available      = available_list[rxn_id] # list[str], list of reactants for this rxn
-                available      = [bb_dict[smiles] for smiles in available] # list[int]
-                temp_emb       = bb_emb[available]
-                available_tree = BallTree(temp_emb, metric=cosine_distance) # TODO: evaluate if distance matrix is faster/feasible as this BallTree is discarded immediately.
-                dist, ind      = nn_search(z_mol2, _tree=available_tree)
-                mol2           = building_blocks[available[ind]]
+                z_mol2 = z_mol2.detach().numpy()
+                available = available_list[rxn_id]  # list[str], list of reactants for this rxn
+                available = [bb_dict[smiles] for smiles in available]  # list[int]
+                temp_emb = bb_emb[available]
+                available_tree = BallTree(
+                    temp_emb, metric=cosine_distance
+                )  # TODO: evaluate if distance matrix is faster/feasible as this BallTree is discarded immediately.
+                dist, ind = nn_search(z_mol2, _tree=available_tree)
+                mol2 = building_blocks[available[ind]]
         else:
             mol2 = None
 
@@ -950,20 +483,23 @@ def synthetic_tree_decoder_rt1(z_target,
 
     return tree, act
 
-def synthetic_tree_decoder_multireactant(z_target,
-                                         building_blocks,
-                                         bb_dict,
-                                         reaction_templates,
-                                         mol_embedder,
-                                         action_net,
-                                         reactant1_net,
-                                         rxn_net,
-                                         reactant2_net,
-                                         bb_emb,
-                                         rxn_template,
-                                         n_bits,
-                                         beam_width : int=3,
-                                         max_step : int=15):
+
+def synthetic_tree_decoder_multireactant(
+    z_target,
+    building_blocks,
+    bb_dict,
+    reaction_templates,
+    mol_embedder,
+    action_net,
+    reactant1_net,
+    rxn_net,
+    reactant2_net,
+    bb_emb,
+    rxn_template,
+    n_bits,
+    beam_width: int = 3,
+    max_step: int = 15,
+):
     """
     Computes the synthetic tree given an input molecule embedding, using the
     Action, Reaction, Reactant1, and Reactant2 networks and a greedy search.
@@ -994,24 +530,27 @@ def synthetic_tree_decoder_multireactant(z_target,
     acts = []
 
     for i in range(beam_width):
-        tree, act = synthetic_tree_decoder_rt1(z_target=z_target,
-                                               building_blocks=building_blocks,
-                                               bb_dict=bb_dict,
-                                               reaction_templates=reaction_templates,
-                                               mol_embedder=mol_embedder,
-                                               action_net=action_net,
-                                               reactant1_net=reactant1_net,
-                                               rxn_net=rxn_net,
-                                               reactant2_net=reactant2_net,
-                                               bb_emb=bb_emb,
-                                               rxn_template=rxn_template,
-                                               n_bits=n_bits,
-                                               max_step=max_step,
-                                               rt1_index=i)
+        tree, act = synthetic_tree_decoder_rt1(
+            z_target=z_target,
+            building_blocks=building_blocks,
+            bb_dict=bb_dict,
+            reaction_templates=reaction_templates,
+            mol_embedder=mol_embedder,
+            action_net=action_net,
+            reactant1_net=reactant1_net,
+            rxn_net=rxn_net,
+            reactant2_net=reactant2_net,
+            bb_emb=bb_emb,
+            rxn_template=rxn_template,
+            n_bits=n_bits,
+            max_step=max_step,
+            rt1_index=i,
+        )
 
-
-        similarities_ = np.array(tanimoto_similarity(z_target, [node.smiles for node in tree.chemicals]))
-        max_simi_idx  = np.where(similarities_ == np.max(similarities_))[0][0]
+        similarities_ = np.array(
+            tanimoto_similarity(z_target, [node.smiles for node in tree.chemicals])
+        )
+        max_simi_idx = np.where(similarities_ == np.max(similarities_))[0][0]
 
         similarities.append(np.max(similarities_))
         smiles.append(tree.chemicals[max_simi_idx].smiles)
@@ -1019,54 +558,19 @@ def synthetic_tree_decoder_multireactant(z_target,
         acts.append(act)
 
     max_simi_idx = np.where(similarities == np.max(similarities))[0][0]
-    similarity   = similarities[max_simi_idx]
-    tree         = trees[max_simi_idx]
-    smi          = smiles[max_simi_idx]
-    act          = acts[max_simi_idx]
+    similarity = similarities[max_simi_idx]
+    tree = trees[max_simi_idx]
+    smi = smiles[max_simi_idx]
+    act = acts[max_simi_idx]
 
     return smi, similarity, tree, act
 
-def fp_embedding(smi, _radius=2, _nBits=4096):
-    """
-    General function for building variable-size & -radius Morgan fingerprints.
-
-    Args:
-        smi (str): The SMILES to encode.
-        _radius (int, optional): Morgan fingerprint radius. Defaults to 2.
-        _nBits (int, optional): Morgan fingerprint length. Defaults to 4096.
-
-    Returns:
-        np.ndarray: A Morgan fingerprint generated using the specified parameters.
-    """
-    if smi is None:
-        return np.zeros(_nBits).reshape((-1, )).tolist()
-    else:
-        mol = Chem.MolFromSmiles(smi)
-        features_vec = AllChem.GetMorganFingerprintAsBitVect(mol, _radius, _nBits)
-        features = np.zeros((1,))
-        DataStructs.ConvertToNumpyArray(features_vec, features)
-        return features.reshape((-1, )).tolist()
-
-def fp_4096(smi):
-    return fp_embedding(smi, _radius=2, _nBits=4096)
-
-def fp_2048(smi):
-    return fp_embedding(smi, _radius=2, _nBits=2048)
-
-def fp_1024(smi):
-    return fp_embedding(smi, _radius=2, _nBits=1024)
-
-def fp_512(smi):
-    return fp_embedding(smi, _radius=2, _nBits=512)
-
-def fp_256(smi):
-    return fp_embedding(smi, _radius=2, _nBits=256)
 
 def rdkit2d_embedding(smi):
     # define the RDKit 2D descriptors conversion function
-    rdkit2d = MolConvert(src = 'SMILES', dst = 'RDKit2D')
+    rdkit2d = MolConvert(src="SMILES", dst="RDKit2D")
 
     if smi is None:
-        return np.zeros(200).reshape((-1, )).tolist()
+        return np.zeros(200).reshape((-1,)).tolist()
     else:
         return rdkit2d(smi).tolist()
